breaking map (warp3d)
